Define the countdowns as Int64

+++ DiaBLE Playground.swiftpm/Console.swift
@@ -302,8 +302,8 @@ struct ConsoleSidebar: View, LoggingView {
 
     @Binding var showingNFCAlert: Bool
 
-    @State private var onlineCountdown: Int = 0
-    @State private var readingCountdown: Int = 0
+    @State private var onlineCountdown: Int64 = 0
+    @State private var readingCountdown: Int64 = 0
 
     let timer = Timer.publish(every: 1, on: .main, in: .common).autoconnect()
 
@@ -374,7 +374,7 @@ struct ConsoleSidebar: View, LoggingView {
                     .font(.caption.monospacedDigit())
                     .foregroundStyle(.orange)
                     .onReceive(timer) { _ in
-                        readingCountdown = settings.readingInterval * 60 - Int(Date().timeIntervalSince(app.lastConnectionDate))
+                        readingCountdown = Int64(settings.readingInterval * 60) - Int64(Date().timeIntervalSince(app.lastConnectionDate))
                     }
                 } else {
                     Text("")
@@ -388,7 +388,7 @@ struct ConsoleSidebar: View, LoggingView {
                     .foregroundStyle(.cyan)
                     .font(.caption.monospacedDigit())
                     .onReceive(timer) { _ in
-                        onlineCountdown = settings.onlineInterval * 60 - Int(Date().timeIntervalSince(settings.lastOnlineDate))
+                        onlineCountdown = Int64(settings.onlineInterval * 60) - Int64(Date().timeIntervalSince(settings.lastOnlineDate))
                     }
 
             }
