+++ DiaBLE/Views/OnlineView.swift
@@ -22,8 +22,8 @@ struct OnlineView: View, LoggingView {
     @Environment(\.colorScheme) var colorScheme
 
     @State private var showingNFCAlert = false
-    @State private var onlineCountdown: Int = 0
-    @State private var readingCountdown: Int = 0
+    @State private var onlineCountdown: Int64 = 0
+    @State private var readingCountdown: Int64 = 0
 
     @State private var timer = Timer.publish(every: 1, on: .main, in: .common).autoconnect()
     @State private var minuteTimer = Timer.publish(every: 60, on: .main, in: .common).autoconnect()
@@ -122,7 +122,7 @@ struct OnlineView: View, LoggingView {
                                 .foregroundStyle(.cyan)
                                 .font(.caption.monospacedDigit())
                                 .onReceive(timer) { _ in
-                                    onlineCountdown = settings.onlineInterval * 60 - Int(Date().timeIntervalSince(settings.lastOnlineDate))
+                                    onlineCountdown = Int64(settings.onlineInterval * 60) - Int64(Date().timeIntervalSince(settings.lastOnlineDate))
                                 }
                         }
 
@@ -142,7 +142,7 @@ struct OnlineView: View, LoggingView {
                             .foregroundStyle(.orange)
                             .font(.caption.monospacedDigit())
                             .onReceive(timer) { _ in
-                                readingCountdown = settings.readingInterval * 60 - Int(Date().timeIntervalSince(app.lastConnectionDate))
+                                readingCountdown = Int64(settings.readingInterval * 60) - Int64(Date().timeIntervalSince(app.lastConnectionDate))
                             }
                         }
